Changed the renderer to use the new RenderPipeline system.
The forward and picking passes are now contained in their own pipelines, since render targets are now stored in
 pipelines to permit sharing between render passes (ie. the target of one pass can be the input of another). Every
render pass must have a pipeline as a parent (this will be moved out of the 'opt' parameter soon).

The picking passes are also responsible for creating their picking targets. Later on we could create a special PickingRenderPipeline
 that inherits from RenderPipeline.