Added named pipelines to make initialization less confusing.
A render pass now expects a pipeline and the name of the output buffer that it should draw to. This buffer can be created in the pipeline's init()
function. To provide a custom pipeline the user would create his own that inherits from XML3D.webgl.RenderPipeline and
override the init/render functions (if necessary). This should make render passes more flexible.